requirements.txt
# Base requirements for the MMIU LLaVA Evaluation Pipeline

# For PyTorch, choose the version appropriate for your target environment (CPU or GPU)
# Example for CPU-only (smaller, but LLaVA inference will be very slow):
# torch torchvision torchaudio
# Example for CUDA 11.8 (check https://pytorch.org/get-started/previous-versions/ for others):
# torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# Example for CUDA 12.1:

# For Hugging Face Transformers and LLaVA models
transformers>=4.38.0 # Ensure a recent version for LLaVA compatibility

# Image processing
Pillow>=9.0.0

# OpenAI API for choice conversion
openai>=1.0.0

# For loading .env files
python-dotenv>=1.0.0

# For progress bars
tqdm>=4.60.0

# Optional: For 8-bit quantization with --use_quantization
# bitsandbytes is primarily for Linux with NVIDIA GPUs.
# datasets
datasets>=2.14.0  # For loading and processing datasets
torch~=2.7.0
peft~=0.15.2
bitsandbytes>=0.41.1  # Uncomment if you want to use 8-bit quantization

# openai tenserboardx 
tensorboardX>=2.6.2  # For logging and visualization
# For evaluation metrics
scikit-learn>=1.0.0  # For evaluation metrics like accuracy, precision